Reverted scripts/do_mv.sh and added scripts/whitespace.sed
Simpler is better.

+++ scripts/do_mv.sh
@@ -1,13 +1,4 @@
 #!/bin/sh
 
-# If they differ, copy the sedded version over the original file, otherwise
-# delete the .sedded version.
-
-if [ -f $1.sedded ]; then
-	if diff $1 $1.sedded > /dev/null; then
-		rm $1.sedded
-	else
-		mv $1.sedded $1
-	fi
-fi
-
+# copy the sedded version over the original file
+[ -f $1.sedded ] && mv $1.sedded $1
